+++ scanner/plugins/javasecurity/javasecurity_test.go
@@ -17,12 +17,15 @@
 package javasecurity
 
 import (
+	"fmt"
+	"github.com/magiconair/properties"
+	"github.com/stretchr/testify/assert"
 	"testing"
 
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 )
 
-func Test_getJDKPath(t *testing.T) {
+func TestJDKPath(t *testing.T) {
 	type args struct {
 		dockerConfig v1.Config
 	}
@@ -85,4 +88,125 @@ func Test_getJDKPath(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestExtractTLSRules(t *testing.T) {
+	t.Run("Extracting TLS Rules from security file", func(t *testing.T) {
+		props := properties.NewProperties()
+		_, _, err := props.Set("jdk.tls.disabledAlgorithms", "SHA384, RSA keySize == 3")
+		if err != nil {
+			t.Error("could not prepare java.security file")
+		}
+
+		javaSecurity := New(*props, "java.security")
+		restrictions, err := javaSecurity.extractTLSRules()
+		javaSecurity.tlsDisabledAlgorithms = restrictions
+
+		assert.NoError(t, err)
+		assert.Len(t, javaSecurity.tlsDisabledAlgorithms, 2)
+		for _, res := range javaSecurity.tlsDisabledAlgorithms {
+			switch res.name {
+			case "RSA":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorEqual)
+				assert.Equal(t, res.keySize, 3)
+			case "SHA384":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorNone)
+				assert.Equal(t, res.keySize, 0)
+			default:
+				assert.FailNow(t, fmt.Sprintf("%v is not a possible algo name", res.name))
+			}
+		}
+	})
+}
+
+func TestExtractTLSRulesNotSupported(t *testing.T) {
+	t.Run("Extracting TLS Rules from security file", func(t *testing.T) {
+		props := properties.NewProperties()
+		_, _, err := props.Set("jdk.tls.disabledAlgorithms", "SHA384, RSA jdkCA")
+		if err != nil {
+			t.Error("could not prepare java.security file")
+		}
+
+		javaSecurity := New(*props, "java.security")
+		restrictions, err := javaSecurity.extractTLSRules()
+		javaSecurity.tlsDisabledAlgorithms = restrictions
+
+		assert.NoError(t, err)
+		assert.Len(t, javaSecurity.tlsDisabledAlgorithms, 1)
+		for _, res := range javaSecurity.tlsDisabledAlgorithms {
+			switch res.name {
+			case "SHA384":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorNone)
+				assert.Equal(t, res.keySize, 0)
+			default:
+				assert.FailNow(t, fmt.Sprintf("%v is not a possible algo name", res.name))
+			}
+		}
+	})
+}
+
+func TestExtractTLSRulesIllegalValue1(t *testing.T) {
+	t.Run("Extracting TLS Rules from security file", func(t *testing.T) {
+		props := properties.NewProperties()
+		_, _, err := props.Set("jdk.tls.disabledAlgorithms", "SHA384, RSA keySize keySize")
+		if err != nil {
+			t.Error("could not prepare java.security file")
+		}
+
+		javaSecurity := New(*props, "java.security")
+		restrictions, err := javaSecurity.extractTLSRules()
+		javaSecurity.tlsDisabledAlgorithms = restrictions
+
+		assert.NoError(t, err)
+		assert.Len(t, javaSecurity.tlsDisabledAlgorithms, 1)
+	})
+}
+
+func TestExtractTLSRulesIllegalValue2(t *testing.T) {
+	t.Run("Extracting TLS Rules from security file", func(t *testing.T) {
+		props := properties.NewProperties()
+		_, _, err := props.Set("jdk.tls.disabledAlgorithms", "SHA384, RSA keySize | 234")
+		if err != nil {
+			t.Error("could not prepare java.security file")
+		}
+
+		javaSecurity := New(*props, "java.security")
+		restrictions, err := javaSecurity.extractTLSRules()
+		javaSecurity.tlsDisabledAlgorithms = restrictions
+
+		assert.NoError(t, err)
+		assert.Len(t, javaSecurity.tlsDisabledAlgorithms, 1)
+	})
+}
+
+func TestExtractTLSRulesInclude(t *testing.T) {
+	t.Run("Extracting TLS Rules from security file", func(t *testing.T) {
+		props := properties.NewProperties()
+		_, _, err := props.Set("jdk.tls.disabledAlgorithms", "SHA384, RSA keySize == 3, DIDYOUGETME keySize >= 123")
+		if err != nil {
+			t.Error("could not prepare java.security file")
+		}
+
+		javaSecurity := New(*props, "java.security")
+		restrictions, err := javaSecurity.extractTLSRules()
+		javaSecurity.tlsDisabledAlgorithms = restrictions
+
+		assert.NoError(t, err)
+		assert.Len(t, javaSecurity.tlsDisabledAlgorithms, 3)
+		for _, res := range javaSecurity.tlsDisabledAlgorithms {
+			switch res.name {
+			case "RSA":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorEqual)
+				assert.Equal(t, res.keySize, 3)
+			case "SHA384":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorNone)
+				assert.Equal(t, res.keySize, 0)
+			case "DIDYOUGETME":
+				assert.Equal(t, res.keySizeOperator, keySizeOperatorGreaterEqual)
+				assert.Equal(t, res.keySize, 123)
+			default:
+				assert.FailNow(t, fmt.Sprintf("%v is not a possible algo name", res.name))
+			}
+		}
+	})
 }
